*: Fix deploy manifests + commit a bundle
Coming back to this operator after a bit, and I noticed not a `make
deploy` essentially resulted in some broken manifest, and only `make
deploy-local` was working, but with self-signed certs.

This commit fixes the kustomize config + certain YAML to ensure make
deploy now work as we expect it to, provided cert-manager is already
present.

It also commits a bundle to the repo, for easy single command installs.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>